Fix some lints and move stuff into own file
